backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.3
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.0
python-multipart==0.0.9
websockets==12.0
Pillow==10.3.0
aiofiles==23.2.1
pydantic==2.7.0
pydantic-settings==2.2.1
gunicorn==21.2.0
